requirement.txt
# requirements.txt
google-cloud-bigquery
streamlit
